pyproject.toml
[project]
name = "devops-engineer-from-scratch-project-313"
version = "0.1.0"
description = "Hexlet DevOps project"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.2",
    "sqlmodel>=0.0.16",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
]

[tool.pytest.ini_options]
testpaths = ["."]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v"

